Remove nested classes
**Why**: Nested classes can make autoloader act up. I am thinking these nested classes are the reason for the "can't insert key into hash during iteration" errors we see on startup
