+++ app/services/encryption/uak_password_verifier.rb
@@ -1,38 +1,10 @@
 module Encryption
   class UakPasswordVerifier
-    PasswordDigest = Struct.new(
-      :encrypted_password,
-      :encryption_key,
-      :password_salt,
-      :password_cost,
-    ) do
-      def self.parse_from_string(digest_string)
-        data = JSON.parse(digest_string, symbolize_names: true)
-        new(
-          data[:encrypted_password],
-          data[:encryption_key],
-          data[:password_salt],
-          data[:password_cost],
-        )
-      rescue JSON::ParserError, TypeError
-        raise EncryptionError, 'digest contains invalid json'
-      end
-
-      def to_s
-        {
-          encrypted_password: encrypted_password,
-          encryption_key: encryption_key,
-          password_salt: password_salt,
-          password_cost: password_cost,
-        }.to_json
-      end
-    end
-
     def self.digest(password)
       salt = SecureRandom.hex(32)
       uak = UserAccessKey.new(password: password, salt: salt)
       uak.build
-      PasswordDigest.new(
+      UakPasswordDigest.new(
         uak.encrypted_password,
         uak.encryption_key,
         salt,
@@ -42,7 +14,7 @@ module Encryption
 
     def self.verify(password:, digest:)
       return false if password.blank?
-      parsed_digest = PasswordDigest.parse_from_string(digest)
+      parsed_digest = UakPasswordDigest.parse_from_string(digest)
       uak = UserAccessKey.new(password: password,
                               salt: parsed_digest.password_salt,
                               cost: parsed_digest.password_cost)
