+++ app/services/job_runner/health_checker_summary.rb
@@ -1,0 +1,9 @@
+module JobRunner
+  HealthCheckerSummary = Struct.new(:healthy, :result) do
+    def as_json(*args)
+      to_h.as_json(*args)
+    end
+
+    alias_method :healthy?, :healthy
+  end
+end
