--- app/services/null_twilio_client.rb
@@ -1,19 +1,0 @@
-class NullTwilioClient
-  HttpClient = Struct.new(:adapter)
-
-  def messages
-    self
-  end
-
-  def calls
-    self
-  end
-
-  def create(_params)
-    # noop
-  end
-
-  def http_client
-    HttpClient.new(adapter: 'foo')
-  end
-end
