+++ spec/controllers/health/job_controller_spec.rb
@@ -9,7 +9,7 @@ RSpec.describe Health::JobsController do
     context 'when jobs are healthy' do
       it 'returns healthy' do
         allow(JobRunner::HealthChecker).to receive(:check).
-          and_return(JobRunner::HealthChecker::Summary.new(true, { 'foo' => true }))
+          and_return(JobRunner::HealthCheckerSummary.new(true, { 'foo' => true }))
 
         action
 
@@ -25,7 +25,7 @@ RSpec.describe Health::JobsController do
     context 'when jobs are unhealthy' do
       before do
         expect(JobRunner::HealthChecker).to receive(:check).
-          and_return(JobRunner::HealthChecker::Summary.new(false, { 'foo' => false }))
+          and_return(JobRunner::HealthCheckerSummary.new(false, { 'foo' => false }))
       end
 
       it 'is a 500' do
