+++ spec/services/encryption/uak_password_verifier_spec.rb
@@ -75,7 +75,7 @@ describe Encryption::UakPasswordVerifier do
 
   it 'raises an encryption error when the password digest is nil' do
     expect do
-      Encryption::UakPasswordVerifier::PasswordDigest.parse_from_string(nil)
+      Encryption::UakPasswordDigest.parse_from_string(nil)
     end.to raise_error(Encryption::EncryptionError)
   end
 end
